+++ app/src/main/java/com/example/karaoke_note/AppBar.kt
@@ -1,12 +1,5 @@
 package com.example.karaoke_note
 
-import android.app.Activity
-import android.content.Context
-import android.content.Intent
-import android.provider.DocumentsContract
-import android.util.Log
-import androidx.activity.compose.rememberLauncherForActivityResult
-import androidx.activity.result.contract.ActivityResultContracts
 import androidx.compose.foundation.border
 import androidx.compose.foundation.horizontalScroll
 import androidx.compose.foundation.layout.Arrangement
@@ -27,13 +20,10 @@ import androidx.compose.material.icons.filled.Close
 import androidx.compose.material.icons.filled.Done
 import androidx.compose.material.icons.filled.FilterAlt
 import androidx.compose.material.icons.filled.Games
-import androidx.compose.material.icons.filled.MoreVert
 import androidx.compose.material.icons.filled.Search
+import androidx.compose.material.icons.filled.Settings
 import androidx.compose.material.icons.outlined.FilterAlt
-import androidx.compose.material3.AlertDialog
 import androidx.compose.material3.BottomSheetDefaults
-import androidx.compose.material3.DropdownMenu
-import androidx.compose.material3.DropdownMenuItem
 import androidx.compose.material3.ExperimentalMaterial3Api
 import androidx.compose.material3.FilterChip
 import androidx.compose.material3.FilterChipDefaults
@@ -43,7 +33,6 @@ import androidx.compose.material3.IconButton
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.ModalBottomSheet
 import androidx.compose.material3.Text
-import androidx.compose.material3.TextButton
 import androidx.compose.material3.TextFieldDefaults
 import androidx.compose.material3.TopAppBar
 import androidx.compose.material3.TopAppBarDefaults
@@ -66,27 +55,12 @@ import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.text.input.ImeAction
 import androidx.compose.ui.unit.dp
 import androidx.navigation.NavController
-import com.example.karaoke_note.data.Artist
 import com.example.karaoke_note.data.ArtistDao
-import com.example.karaoke_note.data.DATABASE_VERSION
 import com.example.karaoke_note.data.FilterSetting
 import com.example.karaoke_note.data.GameKind
-import com.example.karaoke_note.data.Song
 import com.example.karaoke_note.data.SongDao
-import com.example.karaoke_note.data.SongScore
 import com.example.karaoke_note.data.SongScoreDao
 import com.example.karaoke_note.ui.component.CustomTextField
-import com.google.gson.Gson
-import com.google.gson.GsonBuilder
-import com.google.gson.JsonDeserializer
-import com.google.gson.JsonSerializer
-import java.io.BufferedWriter
-import java.io.OutputStreamWriter
-import java.text.SimpleDateFormat
-import java.time.LocalDate
-import java.time.format.DateTimeFormatter
-import java.util.Date
-import java.util.Locale
 
 
 @OptIn(ExperimentalMaterial3Api::class)
@@ -102,8 +76,8 @@ fun AppBar(
     focusManagerOfSearchBar: FocusManager
 ) {
     val canPop = remember { mutableStateOf(false) }
-    val showMenu = remember { mutableStateOf(false) }
     var showSheet by remember { mutableStateOf(false) }
+    var showSettings by remember { mutableStateOf(false) }
     val sheetState = rememberModalBottomSheetState()
 
     LaunchedEffect(navController) {
@@ -208,14 +182,15 @@ fun AppBar(
                     IconButton(
                         onClick = {
                             clearFocusFromSearchBar(focusManagerOfSearchBar)
-                            showMenu.value = true
+                            showSettings = true
                         }
                     ) {
                         Icon(
-                            imageVector = Icons.Filled.MoreVert,
-                            contentDescription = "メニュー"
+                            imageVector = Icons.Filled.Settings,
+                            contentDescription = "Settings"
                         )
                     }
+                    /*
                     DropdownMenu(
                         expanded = showMenu.value,
                         onDismissRequest = { showMenu.value = false }
@@ -227,6 +202,7 @@ fun AppBar(
                             showMenu.value = false
                         }
                     }
+                    */
                 }
             }
         },
@@ -234,6 +210,7 @@ fun AppBar(
         scrollBehavior = TopAppBarDefaults.pinnedScrollBehavior()
     )
 
+    // 下から上がってくるフィルター用画面
     if (showSheet) {
         ModalBottomSheet(
             onDismissRequest = { showSheet = false },
@@ -248,6 +225,13 @@ fun AppBar(
         ) {
             // Sheet content
             FilterContents(filterSetting)
+        }
+    }
+
+    // 設定画面
+    if (showSettings) {
+        SettingScreen(navController, songDao, songScoreDao,artistDao) {
+            showSettings = false
         }
     }
 }
@@ -359,191 +343,6 @@ fun FilterContent(
             }
         } else {
             null
-        }
-    )
-}
-
-
-private fun generateFileName(): String {
-    val dateFormat = SimpleDateFormat("yyyy-MM-dd", Locale.getDefault())
-    val date = dateFormat.format(Date())
-    return "karaoke_note_backup_$date.json"
-}
-
-@Composable
-fun ExportMenu(
-    songDao: SongDao,
-    songScoreDao: SongScoreDao,
-    artistDao: ArtistDao,
-    context: Context,
-    onClick: () -> Unit = {}
-) {
-    val filePickerLauncher = rememberLauncherForActivityResult(
-        contract = ActivityResultContracts.StartActivityForResult()
-    ) { result ->
-        if (result.resultCode == Activity.RESULT_OK) {
-            val selectedFolderUri = result.data?.data
-            // ログに選択されたファイルのパスを表示
-            Log.d("FolderPicker", "Selected folder: $selectedFolderUri")
-            val songScores = songScoreDao.getAll()
-            val songs = songDao.getAllSongs()
-            val artists = artistDao.getAllArtists()
-
-            // LocalDate型のカスタムシリアライザ
-            val localDateSerializer = JsonSerializer<LocalDate> { src, _, _ ->
-                Gson().toJsonTree(src.format(DateTimeFormatter.ISO_LOCAL_DATE))
-            }
-
-            // Gsonインスタンスの作成
-            val gson = GsonBuilder()
-                .registerTypeAdapter(LocalDate::class.java, localDateSerializer)
-                .create()
-
-            // エクスポートするデータ構造
-            val exportData = mapOf(
-                "version" to DATABASE_VERSION,
-                "songScores" to songScores,
-                "songs" to songs,
-                "artists" to artists
-            )
-            val json = gson.toJson(exportData)
-
-            try {
-                val documentUri = DocumentsContract.buildDocumentUriUsingTree(
-                    selectedFolderUri,
-                    DocumentsContract.getTreeDocumentId(selectedFolderUri)
-                )
-
-                val jsonFileUri = DocumentsContract.createDocument(
-                    context.contentResolver,
-                    documentUri,
-                    "application/json",
-                    generateFileName()
-                )
-                if (jsonFileUri == null) {
-                    Log.e("FolderPicker", "Error creating JSON file")
-                } else {
-                    context.contentResolver.openOutputStream(jsonFileUri).use { outputStream ->
-                        BufferedWriter(OutputStreamWriter(outputStream)).use { writer ->
-                            writer.write(json)
-                        }
-                    }
-                }
-                Log.d("FolderPicker", "JSON file saved successfully")
-            } catch (e: Exception) {
-                Log.e("FolderPicker", "Error saving JSON file", e)
-            }
-        }
-        onClick()
-    }
-    DropdownMenuItem(
-        text = {
-            Text("データのエクスポート")
-        },
-        onClick = {
-            val intent = Intent(Intent.ACTION_OPEN_DOCUMENT_TREE)
-            filePickerLauncher.launch(intent)
-        }
-    )
-}
-
-@Composable
-fun ImportMenu(
-    songDao: SongDao,
-    songScoreDao: SongScoreDao,
-    artistDao: ArtistDao,
-    context: Context,
-    onClick: () -> Unit = {}
-) {
-    val localDateDeserializer = JsonDeserializer { json, _, _ ->
-        LocalDate.parse(json.asJsonPrimitive.asString, DateTimeFormatter.ISO_LOCAL_DATE)
-    }
-
-    val gson = GsonBuilder()
-        .registerTypeAdapter(LocalDate::class.java, localDateDeserializer)
-        .create()
-
-    var showDialog by remember { mutableStateOf(false) }
-    val filePickerLauncher = rememberLauncherForActivityResult(
-        contract = ActivityResultContracts.StartActivityForResult()
-    ) { result ->
-        if (result.resultCode == Activity.RESULT_OK) {
-            val selectedFileUri = result.data?.data
-            // ログに選択されたファイルのパスを表示
-            Log.d("FilePicker", "Selected file: $selectedFileUri")
-            try {
-                context.contentResolver.openInputStream(selectedFileUri!!).use { inputStream ->
-                    val json = inputStream?.bufferedReader().use { it?.readText() }
-                    Log.d("FilePicker", "JSON file loaded successfully")
-                    Log.d("FilePicker", json!!)
-                    data class JsonVersion(
-                        val version: Int
-                    )
-                    val versionInfo = gson.fromJson(json, JsonVersion::class.java)
-                    when (versionInfo.version) {
-                        4 -> {
-                            data class JsonDataV3(
-                                val version: Int,
-                                val songScores: List<SongScore>,
-                                val songs: List<Song>,
-                                val artists: List<Artist>
-                            )
-                            val jsonDataV3 = gson.fromJson(json, JsonDataV3::class.java)
-
-                            // IDが混在するとおかしくなるのでデータベースをクリア
-                            songDao.clearAllSongs()
-                            songScoreDao.clearAllSongScores()
-                            artistDao.clearAllArtists()
-
-                            // データベースにインポート
-                            songDao.insertAll(jsonDataV3.songs)
-                            songScoreDao.insertAll(jsonDataV3.songScores)
-                            artistDao.insertAll(jsonDataV3.artists)
-                        }
-                        else -> {
-                            throw IllegalArgumentException("Unsupported version: ${versionInfo.version}")
-                        }
-                    }
-                }
-            } catch (e: Exception) {
-                Log.e("FilePicker", "Error loading JSON file", e)
-            }
-        }
-        onClick()
-    }
-
-    if (showDialog) {
-        AlertDialog(
-            onDismissRequest = { showDialog = false },
-            title = { Text("確認") },
-            text = { Text("すべてのデータは失われますがよろしいですか？") },
-            confirmButton = {
-                TextButton(
-                    onClick = {
-                        showDialog = false
-                        // ファイルピッカーを起動
-                        val intent = Intent(Intent.ACTION_GET_CONTENT).apply {
-                            type = "application/json"
-                        }
-                        filePickerLauncher.launch(intent)
-                    }
-                ) {
-                    Text("はい")
-                }
-            },
-            dismissButton = {
-                TextButton(onClick = { showDialog = false }) {
-                    Text("いいえ")
-                }
-            }
-        )
-    }
-    DropdownMenuItem(
-        text = {
-            Text("データのインポート")
-        },
-        onClick = {
-            showDialog = true
         }
     )
 }
